add v4l2ctdviewer application example
This is a small application that illustrates how to use the
v4l2ctd driver to acces the CRTC buffer. It is primarily intended
to test v4l2ctd and characterize its performance, but an open
minded user can use it to achieve a function similar to
Xephyr or Xnest (taking advantage of virtual CRTC as a kernel
shortcut that eliminates a lot of data moving) or as a remote
desktop mechanism (hint: ssh -X).

Signed-off-by: Hans Christian Woithe <Hans.Woithe@alcatel-lucent.com>
Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>