pimmgr: allow shortening commands
Instead of comparing the command entered by the user with the
whole command string, we allow the user entry to only match
the command prefix. That way, we can type 'pimmgr inst'
instead of long and typo-prone 'pimmgr instantiate'.

If prefix is not unique, the first one in the command search
list is matched (e.g., 'pimmgr in' matches 'instantiate' command,
so we need 'pimmgr inf' if we want 'info' command).

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>