pimmgr: change pimmgr info to not use nftw to search for PIMS and PCONS
Pimmgr was using nftw to walk the sysfs tree to search for PIMS and
PCONS.  To do so, we were making use of glibc specific features.
Since the more standardized ftw is being marked obsolute, we revert
to using more old fashion (opendir, readdir, etc.) functions.  nftw
also had a tendency to walk deeper into the tree than we really
needed it to.

Signed-off-by: Hans Christian Woithe <hans.woithe@alcatel-lucent.com>