pimmgr: modify to use new interface for ioctls
We now use the ioctl interface struct from the kernel, and that
has been updated, so we update pimmgr to match.

The primary difference is that the kernel now requires a pimid to
instantiate a pcon, rather than the pim name as a string. We get
this pimid from sysfs, and call the ioctl as before.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>