pimmgr: just print error as is
There is no point in translating standard Unix errors
into pimmgr's interpretation. If the same kind of error
code occurs somewhere else in the system call path,
"translating" the error will only be confusing.

It is better to keep errors as they are and explain
in the documentation what is the most likely (but
not certain) cause if the error occurs in pimmgr.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>