tools: consistently use sudo in install targets
We should either use sudo for all install targets and let
an ordinary user (with sudo privileges) run 'make install'
or we use sudo nowhere and expect user to type sudo in shell.
We should not mix two ways of installation.

We opt for the former, so this patch brings all install targets
up to snuff.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>