udev_rules: use new vcrtcm sysfs location
The vcrtcm sysfs root directory has moved.  Although this tool gets
installed into a system directory, and it would be inconvenient to
have to re-install an older version of it in order to test an older
kernel, this new version of the script nevertheless uses *only* the
new location, because it would be bad for this script to change modes
on a nonexistent file.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>